LayoutTests/imported/w3c/web-platform-tests/css/css-viewport/zoom/length-implicit-and-explicit-inheritance-expected.txt


FAIL width: inherit from 100px assert_equals: expected "100px" but got "50px"
FAIL height: inherit from 100px assert_equals: expected "100px" but got "50px"
PASS word-spacing: inherit from 100px
PASS word-spacing: unset from 100px
PASS word-spacing:  from 100px
